inference-update
